UHF-10334: added new bundle class for news article, moved the cache logic from news item to recommendablebase-class
