+++ public/modules/custom/helfi_annif/src/RecommendableBase.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\helfi_annif;
 
+use Drupal\Core\Cache\Cache;
 use Drupal\node\Entity\Node;
 
 /**
@@ -42,6 +43,35 @@ abstract class RecommendableBase extends Node implements RecommendableInterface 
    */
   public function getKeywordFieldName(): string {
     return self::KEYWORDFIELD;
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  public function getCacheTagsToInvalidate(): array {
+    $parentCacheTags = parent::getCacheTagsToInvalidate();
+    if (!$this->hasField(self::getKeywordFieldName())) {
+      return $parentCacheTags;
+    }
+
+    $keywordsCacheTags = $this->getKeywordsCacheTags();
+    return Cache::mergeTags($parentCacheTags, $keywordsCacheTags);
+  }
+
+  /**
+   * Get the cache tags for all of the keywords.
+   *
+   * @return array
+   *   Array of cache tags for keywrods.
+   */
+  protected function getKeywordsCacheTags(): array {
+    $terms = $this->get(self::getKeywordFieldName())->referencedEntities();
+
+    $tags = array_map(
+      fn ($term) => $term->getCacheTags(),
+      $terms
+    );
+    return array_merge(...$tags);
   }
 
 }
