+++ public/modules/custom/helfi_etusivu/helfi_etusivu.module
@@ -14,6 +14,7 @@ use Drupal\Core\Field\FieldStorageDefinitionInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Language\LanguageInterface;
 use Drupal\helfi_annif\TextConverter\Document;
+use Drupal\helfi_etusivu\Entity\Node\NewsArticle;
 use Drupal\helfi_etusivu\Entity\Node\NewsItem;
 use Drupal\helfi_platform_config\DTO\ParagraphTypeCollection;
 use Drupal\node\NodeInterface;
@@ -273,5 +274,8 @@ function helfi_etusivu_node_text_conversion_alter(Document &$document, EntityInt
 function helfi_etusivu_entity_bundle_info_alter(array &$bundles): void {
   if (isset($bundles['node']['news_item'])) {
     $bundles['node']['news_item']['class'] = NewsItem::class;
+  }
+  if (isset($bundles['node']['news_article'])) {
+    $bundles['node']['news_article']['class'] = NewsArticle::class;
   }
 }
