+++ public/modules/custom/helfi_etusivu/src/Entity/Node/NewsArticle.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\helfi_etusivu\Entity\Node;
+
+use Drupal\helfi_annif\RecommendableBase;
+
+/**
+ * A bundle class for NewsItem -node.
+ */
+final class NewsArticle extends RecommendableBase {
+}
